# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/products/{id}": {
      "get": {
        "operationId": "getAdminProductById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "getAdminProductVariantGroups": {
                "operationId": "listAdminProductVariantGroups",
                "parameters": {
                  "productId": "$response.body#/id"
                }
              },
              "getPublicProductVariantGroups": {
                "operationId": "getPublicProductVariantGroups",
                "parameters": {
                  "productId": "$response.body#/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Admin Products"
        ]
      },
      "patch": {
        "description": "Supports optional nested variantGroups[] updates: with id updates existing group/option, without id creates new one",
        "operationId": "updateAdminProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "links": {
              "getUpdatedAdminProduct": {
                "operationId": "getAdminProductById",
                "parameters": {
                  "id": "$response.body#/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update product",
        "tags": [
          "Admin Products"
        ]
      },
      "delete": {
        "operationId": "deleteAdminProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Admin Products"
        ]
      }
    },
    "/api/admin/products/{id}/restore": {
      "patch": {
        "operationId": "restoreAdminProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "getRestoredAdminProduct": {
                "operationId": "getAdminProductById",
                "parameters": {
                  "id": "$response.body#/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Admin Products"
        ]
      }
    },
    "/api/admin/products/{productId}/variant-groups": {
      "get": {
        "operationId": "listAdminProductVariantGroups",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Active status filter",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "getVariantGroupFromList": {
                "operationId": "getAdminProductVariantGroup",
                "parameters": {
                  "productId": "$request.path.productId",
                  "groupId": "$response.body#/0/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminProductVariantGroupDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get variant groups for product (admin)",
        "tags": [
          "Admin Product Variants"
        ]
      },
      "post": {
        "operationId": "createAdminProductVariantGroup",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductVariantGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "links": {
              "getCreatedVariantGroup": {
                "operationId": "getAdminProductVariantGroup",
                "parameters": {
                  "productId": "$request.path.productId",
                  "groupId": "$response.body#/id"
                }
              },
              "updateCreatedVariantGroup": {
                "operationId": "updateAdminProductVariantGroup",
                "parameters": {
                  "productId": "$request.path.productId",
                  "groupId": "$response.body#/id"
                }
              },
              "deleteCreatedVariantGroup": {
                "operationId": "deleteAdminProductVariantGroup",
                "parameters": {
                  "productId": "$request.path.productId",
                  "groupId": "$response.body#/id"
                }
              },
              "restoreCreatedVariantGroup": {
                "operationId": "restoreAdminProductVariantGroup",
                "parameters": {
                  "productId": "$request.path.productId",
                  "groupId": "$response.body#/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductVariantGroupDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create variant group for product (admin)",
        "tags": [
          "Admin Product Variants"
        ]
      }
    },
    "/api/admin/products/{productId}/variant-groups/{groupId}": {
      "get": {
        "operationId": "getAdminProductVariantGroup",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "getPublicVariantGroupsForProduct": {
                "operationId": "getPublicProductVariantGroups",
                "parameters": {
                  "productId": "$request.path.productId"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductVariantGroupDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one variant group (admin)",
        "tags": [
          "Admin Product Variants"
        ]
      },
      "patch": {
        "operationId": "updateAdminProductVariantGroup",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductVariantGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "links": {
              "getUpdatedVariantGroup": {
                "operationId": "getAdminProductVariantGroup",
                "parameters": {
                  "productId": "$request.path.productId",
                  "groupId": "$response.body#/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductVariantGroupDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update variant group (admin)",
        "tags": [
          "Admin Product Variants"
        ]
      },
      "delete": {
        "operationId": "deleteAdminProductVariantGroup",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete variant group (admin)",
        "tags": [
          "Admin Product Variants"
        ]
      }
    },
    "/api/admin/products/{productId}/variant-groups/{groupId}/restore": {
      "patch": {
        "operationId": "restoreAdminProductVariantGroup",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "getRestoredVariantGroup": {
                "operationId": "getAdminProductVariantGroup",
                "parameters": {
                  "productId": "$request.path.productId",
                  "groupId": "$response.body#/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductVariantGroupDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore variant group (admin)",
        "tags": [
          "Admin Product Variants"
        ]
      }
    },
    "/api/admin/products/{productId}/variant-groups/{groupId}/options": {
      "post": {
        "operationId": "createAdminProductVariantOption",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductVariantOptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "links": {
              "getGroupForCreatedVariantOption": {
                "operationId": "getAdminProductVariantGroup",
                "parameters": {
                  "productId": "$request.path.productId",
                  "groupId": "$request.path.groupId"
                }
              },
              "updateCreatedVariantOption": {
                "operationId": "updateAdminProductVariantOption",
                "parameters": {
                  "productId": "$request.path.productId",
                  "groupId": "$request.path.groupId",
                  "optionId": "$response.body#/id"
                }
              },
              "deleteCreatedVariantOption": {
                "operationId": "deleteAdminProductVariantOption",
                "parameters": {
                  "productId": "$request.path.productId",
                  "groupId": "$request.path.groupId",
                  "optionId": "$response.body#/id"
                }
              },
              "restoreCreatedVariantOption": {
                "operationId": "restoreAdminProductVariantOption",
                "parameters": {
                  "productId": "$request.path.productId",
                  "groupId": "$request.path.groupId",
                  "optionId": "$response.body#/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductVariantOptionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create variant option in group (admin)",
        "tags": [
          "Admin Product Variants"
        ]
      }
    },
    "/api/admin/products/{productId}/variant-groups/{groupId}/options/{optionId}": {
      "patch": {
        "operationId": "updateAdminProductVariantOption",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "optionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductVariantOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "links": {
              "getGroupForUpdatedVariantOption": {
                "operationId": "getAdminProductVariantGroup",
                "parameters": {
                  "productId": "$request.path.productId",
                  "groupId": "$request.path.groupId"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductVariantOptionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update variant option (admin)",
        "tags": [
          "Admin Product Variants"
        ]
      },
      "delete": {
        "operationId": "deleteAdminProductVariantOption",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "optionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete variant option (admin)",
        "tags": [
          "Admin Product Variants"
        ]
      }
    },
    "/api/admin/products/{productId}/variant-groups/{groupId}/options/{optionId}/restore": {
      "patch": {
        "operationId": "restoreAdminProductVariantOption",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "optionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "getGroupForRestoredVariantOption": {
                "operationId": "getAdminProductVariantGroup",
                "parameters": {
                  "productId": "$request.path.productId",
                  "groupId": "$request.path.groupId"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductVariantOptionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore variant option (admin)",
        "tags": [
          "Admin Product Variants"
        ]
      }
    },
...
  },
  "info": {
    "title": "catalog-service API",
    "description": "REST API + GRPC endpoints",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3002"
    }
  ],
  "components": {
    "schemas": {
...
      "ImageProductDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "Image URL",
            "example": "https://cdn.example.com/products/oak-dining-table-120/main.jpg"
          },
          "type": {
            "type": "string",
            "description": "Image type",
            "example": "main"
          }
        },
        "required": [
          "url",
          "type"
        ]
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Product id",
            "example": 101
          },
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Oak Dining Table 120"
          },
          "fullName": {
            "type": "string",
            "description": "Full product name",
            "example": "Solid Oak Dining Table 120 x 80 cm"
          },
          "sku": {
            "type": "string",
            "description": "SKU (article)",
            "example": "TBL-OAK-120"
          },
          "slug": {
            "type": "string",
            "description": "Product slug",
            "example": "oak-dining-table-120"
          },
          "description": {
            "type": "string",
            "description": "Product description",
            "example": "Natural solid oak dining table with matte finish"
          },
          "price": {
            "type": "number",
            "description": "Product price",
            "example": 24990
          },
          "categoryId": {
            "type": "number",
            "description": "Category id",
            "example": 3
          },
          "categoryName": {
            "type": "string",
            "description": "Category name",
            "example": "Dining Tables"
          },
          "sortOrder": {
            "type": "number",
            "description": "Sorting priority",
            "example": 10
          },
          "images": {
            "example": [
              {
                "url": "https://cdn.example.com/products/oak-dining-table-120/main.jpg",
                "type": "main"
              },
              {
                "url": "https://cdn.example.com/products/oak-dining-table-120/icon.jpg",
                "type": "icon"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageProductDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "price",
          "categoryId",
          "categoryName",
          "sortOrder",
          "images"
        ]
      },
...
    }
  }
}